Fat Package support
This attempts to make binary packages portable over small differences in the OS
execution environment. We have some images based on rocker:shiny and others on
rocker:rstudio and sometimes the rstudio image won't have a library that the
shiny one does. This bundles binary dependencies (but excludes common linux
ones) with the cached R package

This is inspired by:

* [skinnywhale](https://github.com/djosephsen/skinnywhale)
* [conda's way of providing shared
  libraries](https://conda.io/docs/user-guide/tasks/build-packages/use-shared-libraries.html)
* [nixos](https://nixos.wiki/wiki/Packaging/Binaries)